requirements.txt
openai==1.96.1
tweepy==4.16.0
transformers==4.53.2
nltk==3.9.1
vaderSentiment==3.3.2
pandas==2.3.1
requests==2.32.4
python-dotenv==1.0.1
tqdm==4.67.1
torch
PyYAML